+++ Sources/Datadog/Logger.swift
@@ -307,6 +307,7 @@ public class Logger {
         internal var bundleWithTrace = true
         internal var useFileOutput = true
         internal var useConsoleLogFormat: ConsoleLogFormat?
+        internal var datadogReportingThreshold: LogLevel = .debug
 
         /// Sets the service name that will appear in logs.
         /// - Parameter serviceName: the service name  (default value is set to application bundle identifier)
@@ -355,6 +356,14 @@ public class Logger {
         /// - Parameter enabled: `true` by default
         public func sendLogsToDatadog(_ enabled: Bool) -> Builder {
             self.useFileOutput = enabled
+            return self
+        }
+
+        /// Set the minim log level reported to Datadog servers.
+        /// Any log with a level equal or above to the threshold will be sent
+        /// - Parameter datadogReportingThreshold: `LogLevel.debug` by default
+        public func set(datadogReportingThreshold: LogLevel) -> Builder {
+            self.datadogReportingThreshold = datadogReportingThreshold
             return self
         }
 
@@ -437,9 +446,12 @@ public class Logger {
             case (true, let format?):
                 let logOutput = CombinedLogOutput(
                     combine: [
-                        LogFileOutput(
-                            fileWriter: loggingFeature.storage.writer,
-                            rumErrorsIntegration: LoggingWithRUMErrorsIntegration()
+                        ConditionalLogOutput(
+                            conditionedOutput: LogFileOutput(
+                                fileWriter: loggingFeature.storage.writer,
+                                rumErrorsIntegration: LoggingWithRUMErrorsIntegration()
+                            ),
+                            condition: reportLogsAbove(threshold: datadogReportingThreshold)
                         ),
                         LogConsoleOutput(
                             format: format,
@@ -449,9 +461,12 @@ public class Logger {
                 )
                 return (logBuilder, logOutput)
             case (true, nil):
-                let logOutput = LogFileOutput(
-                    fileWriter: loggingFeature.storage.writer,
-                    rumErrorsIntegration: LoggingWithRUMErrorsIntegration()
+                let logOutput = ConditionalLogOutput(
+                    conditionedOutput: LogFileOutput(
+                        fileWriter: loggingFeature.storage.writer,
+                        rumErrorsIntegration: LoggingWithRUMErrorsIntegration()
+                    ),
+                    condition: reportLogsAbove(threshold: datadogReportingThreshold)
                 )
                 return (logBuilder, logOutput)
             case (false, let format?):
@@ -468,5 +483,13 @@ public class Logger {
         private func resolveLoggerName(for loggingFeature: LoggingFeature) -> String {
             return loggerName ?? loggingFeature.configuration.common.applicationBundleIdentifier
         }
+    }
+}
+
+internal func reportLogsAbove(threshold: LogLevel?) -> (LogEvent) -> Bool {
+    return { log in
+        let logSeverity = LogLevel(from: log.status)?.rawValue ?? 0
+        let thresholdValue = threshold?.rawValue ?? .max
+        return logSeverity >= thresholdValue
     }
 }